# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Invoice App API",
        "description": "Dokumentasi API untuk aplikasi invoice.",
        "contact": {
            "email": "admin@example.com"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/invoices": {
            "get": {
                "tags": [
                    "Invoices"
                ],
                "summary": "Get a list of invoices for the authenticated user",
                "operationId": "2e947f49543378e8d1ac24fc3ebcdcc7",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Invoice"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Invoices"
                ],
                "summary": "Create a new invoice",
                "operationId": "15892ddc0ad0649e0919d15345232fc5",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "due_date",
                                    "items"
                                ],
                                "properties": {
                                    "due_date": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2023-09-07"
                                    },
                                    "items": {
                                        "type": "array",
                                        "items": {
                                            "required": [
                                                "item_name",
                                                "qty",
                                                "price"
                                            ],
                                            "properties": {
                                                "item_name": {
                                                    "type": "string",
                                                    "example": "Service 1"
                                                },
                                                "qty": {
                                                    "type": "integer",
                                                    "example": 2
                                                },
                                                "price": {
                                                    "type": "number",
                                                    "format": "float",
                                                    "example": 50000
                                                }
                                            },
                                            "type": "object"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Invoice created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Invoice"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/invoices/{invoice}": {
            "get": {
                "tags": [
                    "Invoices"
                ],
                "summary": "Get a specific invoice",
                "operationId": "a81909c52dc527fa7fef63222e1b052e",
                "parameters": [
                    {
                        "name": "invoice",
                        "in": "path",
                        "description": "Invoice ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Invoice"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Invoice not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Invoices"
                ],
                "summary": "Update an existing invoice",
                "operationId": "8af88d60790a99c0f987103936290092",
                "parameters": [
                    {
                        "name": "invoice",
                        "in": "path",
                        "description": "Invoice ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "due_date",
                                    "items"
                                ],
                                "properties": {
                                    "due_date": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2023-10-07"
                                    },
                                    "items": {
                                        "type": "array",
                                        "items": {
                                            "required": [
                                                "item_name",
                                                "qty",
                                                "price"
                                            ],
                                            "properties": {
                                                "item_name": {
                                                    "type": "string",
                                                    "example": "Updated Item"
                                                },
                                                "qty": {
                                                    "type": "integer",
                                                    "example": 3
                                                },
                                                "price": {
                                                    "type": "number",
                                                    "format": "float",
                                                    "example": 100000
                                                }
                                            },
                                            "type": "object"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Invoice updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Invoice"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Invoices"
                ],
                "summary": "Delete an invoice",
                "operationId": "a60bf154d9b3cb9a2886beff7abba378",
                "parameters": [
                    {
                        "name": "invoice",
                        "in": "path",
                        "description": "Invoice ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Invoice deleted successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Invoice deleted"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
            "Invoice": {
                "title": "Invoice",
                "required": [
                    "due_date",
                    "total"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "user_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "invoice_number": {
                        "type": "string",
                        "example": "INV-20230807-0001"
                    },
                    "due_date": {
                        "type": "string",
                        "format": "date",
                        "example": "2023-09-07"
                    },
                    "total": {
                        "type": "number",
                        "format": "float",
                        "example": 175000
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2023-08-07T12:00:00Z"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2023-08-07T12:00:00Z"
                    },
                    "items": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/InvoiceItem"
                        }
                    }
                },
                "type": "object"
            },
            "InvoiceItem": {
                "title": "InvoiceItem",
                "required": [
                    "item_name",
                    "qty",
                    "price"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "invoice_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "item_name": {
                        "type": "string",
                        "example": "Service 1"
                    },
                    "qty": {
                        "type": "integer",
                        "example": 2
                    },
                    "price": {
                        "type": "number",
                        "format": "float",
                        "example": 50000
                    },
                    "subtotal": {
                        "type": "number",
                        "format": "float",
                        "example": 100000
                    }
                },
                "type": "object"
            },
...
        }
    },
...
}